Add test

+++ src/UnitTestsCourse/Mocking/EmployeeStorage.cs
@@ -28,6 +28,10 @@ namespace UnitTestsCourse.Mocking
         public void DeleteEmployee(int id)
         {
             var employee = _db.Employees.Find(id);
+
+            if (employee == null)
+                return;
+
             _db.Employees.Remove(employee);
             _db.SaveChanges();
         }
